runtime: slience warning on 32-bit platforms
LGTM=iant
R=golang-codereviews, iant
CC=golang-codereviews
https://golang.org/cl/80250043

+++ src/pkg/runtime/heapdump.c
@@ -803,7 +803,7 @@ playgcprog(uintptr offset, uintptr *prog, void (*callback)(void*,uintptr,uintptr
 			prog += 4;
 			break;
 		default:
-			runtime·printf("%D\n", prog[0]);
+			runtime·printf("%D\n", (uint64)prog[0]);
 			runtime·throw("bad gc op");
 		}
 	}
